Changes made to drizzle source when building in the events plugin.

+++ drizzled/cursor.cc
@@ -43,6 +43,7 @@
 #include "drizzled/message/table.pb.h"
 #include "drizzled/plugin/client.h"
 #include "drizzled/internal/my_sys.h"
+#include "drizzled/plugin/event.h"
 
 using namespace std;
 
@@ -1481,9 +1482,17 @@ int Cursor::insertRecord(unsigned char *buf)
   if (table->timestamp_field_type & TIMESTAMP_AUTO_SET_ON_INSERT)
     table->timestamp_field->set_time();
 
-  DRIZZLE_INSERT_ROW_START(table_share->getSchemaName(), table_share->getTableName());
+  DRIZZLE_INSERT_ROW_START(table->getSchemaName(), table_share->getTableName());
   setTransactionReadWrite();
-  error= doInsertRecord(buf);
+  
+  if (unlikely(plugin::Event::preWriteRowDo(table->in_use, table_share, buf)))
+    error= ER_EVENT_PLUGIN;
+  else
+  {
+    error= doInsertRecord(buf);
+    plugin::Event::postWriteRowDo(table->in_use, table_share, buf, error);
+  }
+ 
   DRIZZLE_INSERT_ROW_DONE(error);
 
   if (unlikely(error))
@@ -1510,7 +1519,14 @@ int Cursor::updateRecord(const unsigned char *old_data, unsigned char *new_data)
 
   DRIZZLE_UPDATE_ROW_START(table_share->getSchemaName(), table_share->getTableName());
   setTransactionReadWrite();
-  error= doUpdateRecord(old_data, new_data);
+  if (unlikely(plugin::Event::preUpdateRowDo(table->in_use, table_share, old_data, new_data)))
+    error= ER_EVENT_PLUGIN;
+  else
+  {
+    error= doUpdateRecord(old_data, new_data);
+    plugin::Event::postUpdateRowDo(table->in_use, table_share, old_data, new_data, error);
+  }
+
   DRIZZLE_UPDATE_ROW_DONE(error);
 
   if (unlikely(error))
@@ -1530,7 +1546,14 @@ int Cursor::deleteRecord(const unsigned char *buf)
 
   DRIZZLE_DELETE_ROW_START(table_share->getSchemaName(), table_share->getTableName());
   setTransactionReadWrite();
-  error= doDeleteRecord(buf);
+  if (unlikely(plugin::Event::preDeleteRowDo(table->in_use, table_share, buf)))
+    error= ER_EVENT_PLUGIN;
+  else
+  {
+    error= doDeleteRecord(buf);
+    plugin::Event::postDeleteRowDo(table->in_use, table_share, buf, error);
+  }
+
   DRIZZLE_DELETE_ROW_DONE(error);
 
   if (unlikely(error))
